db: never return a nil iterator on success from newRangeDelIter
`compaction.newInputIter` iterates over the inputs to a compaction by
constructing a merging iter composed of 4 level iterators: 1 point and 1
range-del `levelIter` for the two input levels. The range-del
`levelIter` is constructed using a `newRangeDelIter` closure. This
closure was returning `nil` if an sstable did not contain a range
tombstones. Unfortunately, this interacted badly with
`levelIter.loadFile` which had an odd contional to assume that return a
`nil` iterator and a `nil` error indicated some sort of "end of
iteration" indication. This was completely bogus. The interaction of
these two problems was droppage of range tombstones during a compaction
whenever an sstable contained a range tombstone and an earlier sstable
in the same level that was part of the compaction did not contain any
range tombstones. I suspect this is the cause of some of the roachtest
consistency failures we've seen.

Found via the metamorphic test.